task2/src/backend/backend_requirements.txt
pandas==1.4.2
psycopg2-binary
requests